requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
Pillow==10.0.1
python-jose==1.7.0
passlib==1.7.4
python-dotenv==1.0.0
bitsandbytes
